implement decryptAes128Gcm that wraps around openssl
We are running into an issue when decrypting the payload of Initial
packet sent from the existing client implementation, aws/s2n-quic and
cloudflare/quiche. Sometimes decryption is done successfully, but
sometimes fails. We implemented decryptAes128Gcm that makes use of
openssl to see if our issue has something to do with the implementation
of Zig std's AES128-GCM. Unfortunately, it turned out that the issue
still remains even with openssl. We need to look into it further.